homeworks/homework_13/pyproject.toml
[tool.poetry]
name = "homework-13"
version = "0.1.0"
description = ""
authors = ["serzhu <sergej_zhu@ukr.net>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.111.0"
uvicorn = {extras = ["standard"], version = "^0.29.0"}
sqlalchemy = "^2.0.30"
psycopg2 = "^2.9.9"
alembic = "^1.13.1"
passlib = "^1.7.4"
fastapi-jwt-auth = "^0.5.0"
python-jose = "^3.3.0"
python-multipart = "^0.0.9"
pydantic-extra-types = "^2.7.0"
fastapi-mail = "^1.4.1"
fastapi-limiter = "^0.1.6"
asyncpg = "^0.29.0"
phonenumbers = "^8.13.37"
bcrypt = "^4.1.3"
cloudinary = "^1.40.0"
redis = "^5.0.4"


[tool.poetry.group.dev.dependencies]
sphinx = "^7.3.7"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
